Create README.md
Hello, my name is Yashveer Sekhon and this is my final semester project for my Adv. Java class. We got to pick what we what to create, so I wanted to learn how to use A.I. in TicTacToe. In this project, the user gets to choose 2 modes to play. One mode is regular 1v1 and the other mode is the user vs the computer. The A.I. uses MiniMax to play against the user. A minimax algorithm is a recursive program that finds the best gameplay that minimizes any tendency to lose a game while maximizing any opportunity to win the game.  We can represent minimax as an exploration of a game tree's nodes to discover the best game move to make. Thank you and have a wonderful day.